python/requirements.txt
librosa>=0.10.1
pydub>=0.25.1
numpy>=1.26.0
scipy>=1.11.0
python-speech-features>=0.6
arabic-reshaper>=3.0.0
transformers>=4.36.0
torch>=2.1.0
openai>=1.12.0
requests>=2.31.0
praat-parselmouth>=0.4.3
fastdtw>=0.3.4
